// eform-client/src/app/plugins/modules/my-microting-pn/models/droplets/my-microting-pn-droplets-request-model.ts
export class MyMicrotingPnDropletsRequestModel {
  name: string;
  sortColumnName: string;
  pageIndex: number;
  offset: number;
  pageSize: number;
  isSortDsc: boolean;

  // constructor() {
  //   this.sortColumnName = 'Id';
  //   this.isSortDsc = true;
  //   this.pageSize = 5;
  //   this.pageIndex = 1;
  //   this.offset = 0;
  // }
}